requirements.txt
scikit-learn
numpy
pymupdf
putio.py
ebooklib
html2text
PyMySQL
python-magic
tqdm
python-dotenv
pydantic
pydantic-settings